requirements.txt
keyring
pan-os-python
pyyaml